@@ -1,6 +1,6 @@
 {
     "name": "@dinomite-studios/unity-azure-pipelines-tasks-lib",
-    "version": "1.0.4",
+    "version": "1.0.5",
     "description": "A library containing common implementations for pipeline tasks available in the Unity Tools for Azure DevOps extension.",
     "main": "dist/index.js",
     "types": "dist/index.d.ts",
@@ -26,21 +26,21 @@
     "author": "Dinomite",
     "license": "MIT",
     "dependencies": {
-        "azure-pipelines-task-lib": "^4.13.0",
+        "azure-pipelines-task-lib": "^4.17.2",
         "fs-extra": "^11.2.0",
         "sanitize-filename": "^1.6.3",
         "tail": "^2.2.6"
     },
     "devDependencies": {
-        "@types/chai": "^4.3.16",
-        "@types/mocha": "^10.0.6",
-        "@types/node": "^20.14.5",
+        "@types/chai": "^4.3.0",
+        "@types/mocha": "^10.0.8",
+        "@types/node": "^22.5.5",
         "@types/fs-extra": "^11.0.4",
         "@types/tail": "^2.2.3",
         "@types/q": "^1.5.8",
-        "chai": "^5.1.1",
-        "mocha": "^10.4.0",
-        "ts-node": "^10.7.0",
-        "typescript": "^4.6.2"
+        "chai": "^4.5.0",
+        "mocha": "^10.7.3",
+        "ts-node": "^10.9.2",
+        "typescript": "^5.6.2"
     }
 }
